Remove some redundant casts

+++ src/animation.cpp
@@ -6,10 +6,7 @@ Animation::Animation(const std::string& name, const sf::Texture& texture)
 
 Animation::Animation(const std::string& name, const sf::Texture& texture, size_t frame_count, int speed, Vec2 size, Vec2 offset)
 	: m_name(name), m_sprite(texture), m_frame_count(frame_count), m_current_frame(0), m_speed(speed), m_offset(offset) {
-		m_size = Vec2(
-			static_cast<float>(size.x / static_cast<unsigned int>(frame_count)),
-			static_cast<float>(size.y)
-		);
+		m_size = Vec2(size.x / frame_count, size.y);
 		m_sprite.setOrigin(m_size.x / 2.0f, m_size.y / 2.0f);
 		m_sprite.setTextureRect(sf::IntRect(
 			static_cast<int>(m_offset.x) + static_cast<int>(std::floor(m_current_frame)) * static_cast<int>(m_size.x),
